requirements.txt
google-generativeai
python-dotenv
mysql-connector-python